* commit '0a2bb11b248df6841daabca3970df5d8504adfc7':
  address problems from comments
  fix and address comments
  change according to previous comments
  add missing piece
  remove unnecessary check macro
  fix intel compile warnings
  Revert "fix warnings from Intel compiler"
  Revert "fix warnings and some text alignment"
  Revert "let hdf5 pick up the right compiler in Intel environment"
  Revert "fix issues from previous PR comments"
  Revert "using a different MACRO"
  using a different MACRO
  fix issues from previous PR comments
  let hdf5 pick up the right compiler in Intel environment
  fix warnings and some text alignment
  fix warnings from Intel compiler

+++ hl/tools/gif2h5/decompress.c
@@ -173,7 +173,7 @@ Decompress(GIFIMAGEDESC *GifImageDesc, GIFHEAD *GifHead)
     /* Now read in values from the image descriptor */
     IWidth = GifImageDesc->ImageWidth;
     IHeight = GifImageDesc->ImageHeight;
-    Interlace = GifImageDesc->PackedField & 0x40;
+    Interlace = (uint8_t)(GifImageDesc->PackedField & 0x40);
 
     /*
      * Note that I ignore the possible existence of a local color map.  I'm
